chore: app menu refactor
