+++ src/components/app/styles/AppMenu.module.css
@@ -1,0 +1,7 @@
+.appMenu {
+    position: absolute;
+    width: 100%;
+    height: 39px;
+    z-index: 1200;
+    box-shadow: 0 1px 1px 0 #9e9e9e;
+}
